backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.6.0
pymongo==4.9.0
dnspython==2.7.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
groq==0.4.1
python-dotenv==1.0.0
aiofiles==23.2.1